changing to [%lu]

+++ search_algorithms/102-interpolation.c
@@ -26,7 +26,8 @@ int interpolation_search(int *array, size_t size, int value)
 		pos = low + (((double)(high - low) / (array[high] - array[low]))
 				* (value - array[low]));
 
-		printf("Value checked array[%lu] = [%d]\n", pos, array[pos]);
+		printf("Value checked array[%lu] = [%d]\n", (unsigned long)pos,
+				array[pos]);
 
 		if (array[pos] < value)
 			low = pos + 1;
